Rename preamble to prelude to be consistent with other Rust projects
