+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components.rs
@@ -2,7 +2,7 @@ pub mod common;
 pub mod component_creation;
 pub mod doenet;
 pub mod special;
-pub mod preamble;
+pub mod prelude;
 
 pub use common::*;
 pub use special::*;
