+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/doenet/section.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::components::preamble::*;
+use crate::components::prelude::*;
 
 #[derive(Debug, Default, ComponentNode, RenderedComponentNode)]
 pub struct Section {
