+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/doenet/text_input.rs
@@ -4,7 +4,7 @@ use std::rc::Rc;
 
 use serde::{Deserialize, Serialize};
 
-use crate::components::preamble::*;
+use crate::components::prelude::*;
 
 #[derive(Debug, Default, ComponentNode)]
 pub struct TextInput {
