+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/special/_external.rs
@@ -1,4 +1,4 @@
-use crate::components::preamble::*;
+use crate::components::prelude::*;
 
 #[derive(Debug, Default, RenderedComponentNode, ComponentNodeStateVariables)]
 pub struct _External {
